refactored db.entity package
