+++ account/src/main/java/com/microbank/account/infrastructure/db/entity/AccountEntity.java
@@ -1,4 +1,4 @@
-package com.microbank.account.infrastructure.db.repository.entity;
+package com.microbank.account.infrastructure.db.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
