+++ account/src/main/java/com/microbank/account/infrastructure/db/entity/Entity.java
@@ -1,4 +1,4 @@
-package com.microbank.account.infrastructure.db.repository.entity;
+package com.microbank.account.infrastructure.db.entity;
 
 import lombok.Data;
