+++ account/src/main/java/com/microbank/account/infrastructure/db/entity/UserEntity.java
@@ -1,6 +1,5 @@
-package com.microbank.account.infrastructure.db.repository.entity;
+package com.microbank.account.infrastructure.db.entity;
 
-import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
